iou in tf.py_func return float32

+++ keras/metrics.py
@@ -71,7 +71,7 @@ def compute_eval_metric(gt, predictions):
     thresholds = [0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9, 0.95]
     ious = compute_ious(gt, predictions)
     precisions = [compute_precision_at(ious, th) for th in thresholds]
-    return sum(precisions) / len(precisions)
+    return np.float32(sum(precisions) / len(precisions))
 
 def threshold_mean_iou_metric(gt, predictions):
     metric_value = tf.py_func(compute_eval_metric, [gt, predictions], tf.float32)
